+++ src/language-html/clean.js
@@ -1,9 +1,14 @@
 "use strict";
 
 module.exports = function(ast, newNode) {
-  delete newNode.__location;
+  delete newNode.sourceCodeLocation;
 
   if (ast.type === "text") {
+    return null;
+  }
+
+  // may be formatted by multiparser
+  if (ast.type === "yaml") {
     return null;
   }
 };