+++ src/language-html/embed.js
@@ -2,8 +2,8 @@
 
 const { hasNewlineInRange } = require("../common/util");
 const {
-  builders: { hardline, concat },
-  utils: { stripTrailingHardline, removeLines }
+  builders: { hardline, concat, markAsRoot, literalline },
+  utils: { removeLines, mapDoc }
 } = require("../doc");
 
 function embed(path, print, textToDoc, options) {
@@ -41,7 +41,7 @@ function embed(path, print, textToDoc, options) {
       // Inline Styles
       if (parent.type === "style") {
         const doc = textToDoc(getText(options, node), { parser: "css" });
-        return concat([hardline, stripTrailingHardline(doc)]);
+        return concat([hardline, doc]);
       }
 
       break;
@@ -49,7 +49,7 @@ function embed(path, print, textToDoc, options) {
 
     case "attribute": {
       /*
-       * Vue binding sytax: JS expressions
+       * Vue binding syntax: JS expressions
        * :class="{ 'some-key': value }"
        * v-bind:id="'list-' + id"
        * v-if="foo && !bar"
@@ -57,7 +57,7 @@ function embed(path, print, textToDoc, options) {
        */
       if (/(^@)|(^v-)|:/.test(node.key) && !/^\w+$/.test(node.value)) {
         const doc = textToDoc(node.value, {
-          parser: parseJavaScriptExpression,
+          parser: "__js_expression",
           // Use singleQuote since HTML attributes use double-quotes.
           // TODO(azz): We still need to do an entity escape on the attribute.
           singleQuote: true
@@ -71,18 +71,39 @@ function embed(path, print, textToDoc, options) {
           '"'
         ]);
       }
+
+      break;
     }
+
+    case "yaml":
+      return markAsRoot(
+        concat([
+          "---",
+          hardline,
+          node.value.trim()
+            ? replaceNewlinesWithLiterallines(
+                textToDoc(node.value, { parser: "yaml" })
+              )
+            : "",
+          "---",
+          hardline
+        ])
+      );
   }
 }
 
-function parseJavaScriptExpression(text, parsers) {
-  // Force parsing as an expression
-  const ast = parsers.babylon(`(${text})`);
-  // Extract expression from the declaration
-  return {
-    type: "File",
-    program: ast.program.body[0].expression
-  };
+function replaceNewlinesWithLiterallines(doc) {
+  return mapDoc(
+    doc,
+    currentDoc =>
+      typeof currentDoc === "string" && currentDoc.includes("\n")
+        ? concat(
+            currentDoc
+              .split(/(\n)/g)
+              .map((v, i) => (i % 2 === 0 ? v : literalline))
+          )
+        : currentDoc
+  );
 }
 
 function getText(options, node) {