+++ src/language-html/parser-parse5.js
@@ -1,61 +1,76 @@
 "use strict";
 
+const htmlTagNames = require("html-tag-names");
+const parseFrontMatter = require("../utils/front-matter");
+
+const nonFragmentRegex = /^\s*(<!--[\s\S]*?-->\s*)*<(!doctype|html|head|body)[\s>]/i;
+
 function parse(text /*, parsers, opts*/) {
   // Inline the require to avoid loading all the JS if we don't use it
   const parse5 = require("parse5");
-  try {
-    const isFragment = !/^\s*<(!doctype|html|head|body|!--)/i.test(text);
-    const ast = (isFragment ? parse5.parseFragment : parse5.parse)(text, {
-      treeAdapter: parse5.treeAdapters.htmlparser2,
-      locationInfo: true
-    });
-    return normalize(extendAst(ast));
-  } catch (error) {
-    throw error;
-  }
-}
+  const htmlparser2TreeAdapter = require("parse5-htmlparser2-tree-adapter");
 
-function normalize(ast) {
-  if (Array.isArray(ast)) {
-    return ast.map(normalize);
-  }
+  const { frontMatter, content } = parseFrontMatter(text);
 
-  if (!ast || typeof ast !== "object") {
-    return ast;
-  }
-
-  delete ast.parent;
-  delete ast.next;
-  delete ast.prev;
-
-  for (const key of Object.keys(ast)) {
-    ast[key] = normalize(ast[key]);
-  }
-
-  return ast;
-}
-
-function extendAst(ast) {
-  if (!ast || !ast.children) {
-    return ast;
-  }
-  for (const child of ast.children) {
-    extendAst(child);
-    if (child.attribs) {
-      child.attributes = convertAttribs(child.attribs);
-    }
-  }
-  return ast;
-}
-
-function convertAttribs(attribs) {
-  return Object.keys(attribs).map(attributeKey => {
-    return {
-      type: "attribute",
-      key: attributeKey,
-      value: attribs[attributeKey]
-    };
+  const isFragment = !nonFragmentRegex.test(content);
+  const ast = (isFragment ? parse5.parseFragment : parse5.parse)(content, {
+    treeAdapter: htmlparser2TreeAdapter,
+    sourceCodeLocationInfo: true
   });
+
+  const normalizedAst = normalize(ast, text);
+
+  if (frontMatter) {
+    normalizedAst.children.unshift(frontMatter);
+  }
+
+  return normalizedAst;
+}
+
+function normalize(node, text) {
+  delete node.parent;
+  delete node.next;
+  delete node.prev;
+
+  let isCaseSensitiveTag = false;
+
+  // preserve case-sensitive tag names
+  if (
+    node.type === "tag" &&
+    node.sourceCodeLocation &&
+    htmlTagNames.indexOf(node.name) === -1
+  ) {
+    isCaseSensitiveTag = true;
+    node.name = text.slice(
+      node.sourceCodeLocation.startOffset + 1, // <
+      node.sourceCodeLocation.startOffset + 1 + node.name.length
+    );
+  }
+
+  if (node.attribs) {
+    node.attributes = Object.keys(node.attribs).map(attributeKey => {
+      const sourceCodeLocation = node.sourceCodeLocation.attrs[attributeKey];
+      return {
+        type: "attribute",
+        key: isCaseSensitiveTag
+          ? text
+              .slice(
+                sourceCodeLocation.startOffset,
+                sourceCodeLocation.endOffset
+              )
+              .split("=", 1)[0]
+          : attributeKey,
+        value: node.attribs[attributeKey],
+        sourceCodeLocation
+      };
+    });
+  }
+
+  if (node.children) {
+    node.children = node.children.map(child => normalize(child, text));
+  }
+
+  return node;
 }
 
 module.exports = {
@@ -64,10 +79,10 @@ module.exports = {
       parse,
       astFormat: "htmlparser2",
       locStart(node) {
-        return node.__location && node.__location.startOffset;
+        return node.sourceCodeLocation && node.sourceCodeLocation.startOffset;
       },
       locEnd(node) {
-        return node.__location && node.__location.endOffset;
+        return node.sourceCodeLocation && node.sourceCodeLocation.endOffset;
       }
     }
   }