+++ tests/flow_internal_slot/internal_slot.js
@@ -1,0 +1,7 @@
+declare class C { static [[foo]]: T }
+declare class C { [[foo]]: T }
+interface T { [[foo]]: X }
+interface T { [[foo]](): X }
+type T = { [[foo]]: X }
+type T = { [[foo]](): X }
+type T = { [[foo]]?: X }